feat: Moved stonePosition from game-state-dao to stone-handler, removed all canvas sizes from game-state to game-settings, pointed (FM-380).
