+++ src/components/stone-handler/stone-handler.ts
@@ -13,8 +13,10 @@ import {
   AUDIO_PATH_ON_DRAG
 } from '@constants';
 import gameStateService from '@gameStateService';
+import gameSetttingsService from '@gameSettingsService';
 
 export default class StoneHandler extends EventManager {
+  private offsetCoordinateValue: number;
   public context: CanvasRenderingContext2D;
   public canvas: HTMLCanvasElement;
   public currentPuzzleData: any;
@@ -52,12 +54,14 @@ export default class StoneHandler extends EventManager {
       loadPuzzleCallbackHandler: (event) => this.handleLoadPuzzle(event),
     });
     this.cleanup();
+    this.offsetCoordinateValue = 32; //Default value used to offset stone coordinates.
+    const { } = gameSetttingsService.getCanvasSizeValues();
     this.context = context;
     this.canvas = canvas;
     this.puzzleNumber = puzzleNumber;
     this.levelData = levelData;
     this.setTargetStone(this.puzzleNumber);
-    this.stonePos = gameStateService.getStonePositions();
+    this.stonePos = this.getRandomizedStonePositions(canvas.width, canvas.height)
     this.correctStoneAudio = new Audio(AUDIO_PATH_CORRECT_STONE);
     this.correctStoneAudio.loop = false;
     this.feedbackAudios = this.convertFeedBackAudiosToList(feedbackAudios);
@@ -329,7 +333,7 @@ export default class StoneHandler extends EventManager {
     try {
       // Dispose stones immediately - don't wait for audio
       this.disposeStones();
-      
+
       // Play feedback audio in parallel for better performance
       const randomNumber = Utils.getRandomNumber(1, 3).toString();
       await Promise.allSettled([
@@ -353,9 +357,9 @@ export default class StoneHandler extends EventManager {
       this.correctStoneAudio.pause();
       this.correctStoneAudio.src = '';
     }
-    
+
     this.disposeStones();
-    
+
     // Remove event listeners
     document.removeEventListener(VISIBILITY_CHANGE, this.handleVisibilityChange);
     if (this.unsubscribeEvent) {
@@ -449,5 +453,83 @@ export default class StoneHandler extends EventManager {
     if (stone.frame > 99) {
       this.audioPlayer.playAudio(AUDIO_PATH_ON_DRAG);
     }
+  }
+
+  private getRandomizedStonePositions(widthVal, heightVal) {
+
+    /**
+     * Breakpoint width for stone positioning (in pixels).
+     * Standard layout > 540px, compact layout ≤ 540px.
+     * 
+     * @type {number}
+     * @default 540
+     */
+    const deviceWidth = 540;
+
+    /**
+     * Calculates coordinate factors for stone positioning based on screen width.
+     * Temporary solution until Rive animation integration.
+     * 
+     * @param {number} defaultVal - Factor for screens > 540px (2.0-5.0)
+     * @param {number} smallerVal - Factor for screens ≤ 540px (1.2-1.5x larger)
+     * @returns {number} Calculated coordinate factor
+     * 
+     * @todo Replace with Rive-based dynamic positioning
+     */
+    const setCoordinateFactor = (defaultVal, smallerVal) => {
+      return deviceWidth > widthVal ? smallerVal : defaultVal;
+    }
+
+    /*
+    *  If rive width and height are not properly set,
+    *  use the original canvas width and height instead.
+    *  this is the default coordinates
+    */
+    const baseCoordinateFactors = [
+      [5, 1.9], //Left stone 1 - upper
+      [7, 1.5], //Left stone 2
+      [setCoordinateFactor(4.3, 4.5), 1.28], //Left stone 3
+      [6.4, 1.1], //Left stone 4 - very bottom
+      [setCoordinateFactor(2, 1.3), 1.07], //Middle stone that is located right below the monster.
+      [[2.3, 2.1], 1.9], //Right stone 1 - upper
+      [[setCoordinateFactor(2.8, 2.5), 2], 1.2], //Right stone 2
+      [[setCoordinateFactor(3, 2.4), 2.1], 1.42],  //Right stone 3
+    ];
+
+    // Separate coordinate factors for egg monster due to different dimensions
+    const eggMonsterCoordinateFactors = [
+      [5, 1.9], //Left stone 1 - upper
+      [7, 1.5], //Left stone 2
+      [setCoordinateFactor(4.3, 4.5), 2.28], //Left stone 3
+      [6.4, 1.1], //Left stone 4 - very bottom
+      [setCoordinateFactor(1.2, 1.5), 1], //Middle stone that is located right below the monster.
+      [[2.3, 1.9], 1.5], //Right stone 1 - upper
+      [[setCoordinateFactor(2.8, 2.2), 2.1], 2.4], //Right stone 2 - increased horizontal spacing
+      [[setCoordinateFactor(4.5, 3.2), 1.5], 1.8],  //Right stone 3 - moved further right and down
+    ];
+
+    // Choose coordinate factors based on monster type
+    const coordinateFactors = eggMonsterCoordinateFactors;
+
+    const randomizedStonePositions = coordinateFactors.map(
+      (coordinatesFactors: [number[] | number, number], index) => {
+        const factorX = coordinatesFactors[0];
+        const factorY = coordinatesFactors[1];
+        let coordinateX = Array.isArray(factorX)
+          ? ((widthVal / factorX[0]) + (widthVal / factorX[1]))
+          : (widthVal / factorX);
+        let coordinateY = heightVal / factorY;
+        const offsetXAdjustment = index < 4 ? 25 : 0; //Only use +25 on stones on the left side.
+        const posX = coordinateX - this.offsetCoordinateValue;
+        const posY = coordinateY - this.offsetCoordinateValue;
+
+        return [
+          posX + offsetXAdjustment,
+          posY,
+        ]
+      }
+    ).sort(() => Math.random() - 0.5);
+
+    return randomizedStonePositions;
   }
 }
