+++ src/gameStateService/data-access-objects.spec.ts
@@ -1,97 +1,97 @@
-import { createStonePositionsDAO } from './data-access-objects';
+// import { createStonePositionsDAO } from './data-access-objects';
 
-describe('Data Access Objects - Stone Positioning', () => {
-    const mockDimensions = {
-        width: 800,
-        height: 600,
-        monsterType: 'default',
-        riveCanvasWidth: 800,
-        riveCanvasHeight: 600,
-        offsetCoordinateValue: 0
-    };
+// describe('Data Access Objects - Stone Positioning', () => {
+//     const mockDimensions = {
+//         width: 800,
+//         height: 600,
+//         monsterType: 'default',
+//         riveCanvasWidth: 800,
+//         riveCanvasHeight: 600,
+//         offsetCoordinateValue: 0
+//     };
 
-    describe('Stone Position Calculations', () => {
-        it('should return correct coordinates for stones', () => {
-            const positions = createStonePositionsDAO(mockDimensions);
+//     describe('Stone Position Calculations', () => {
+//         it('should return correct coordinates for stones', () => {
+//             const positions = createStonePositionsDAO(mockDimensions);
             
-            // Verify we get the expected number of stone positions
-            expect(positions).toHaveLength(8); // Total number of stones
+//             // Verify we get the expected number of stone positions
+//             expect(positions).toHaveLength(8); // Total number of stones
             
-            // Verify structure of returned positions
-            positions.forEach((position: number[]) => {
-                expect(position).toHaveLength(2); // [x, y] coordinates
-                expect(typeof position[0]).toBe('number'); // x coordinate
-                expect(typeof position[1]).toBe('number'); // y coordinate
-            });
+//             // Verify structure of returned positions
+//             positions.forEach((position: number[]) => {
+//                 expect(position).toHaveLength(2); // [x, y] coordinates
+//                 expect(typeof position[0]).toBe('number'); // x coordinate
+//                 expect(typeof position[1]).toBe('number'); // y coordinate
+//             });
 
-            // Verify stone positions are within canvas bounds
-            positions.forEach((position: number[]) => {
-                const [x, y] = position;
-                expect(x).toBeGreaterThanOrEqual(0);
-                expect(x).toBeLessThanOrEqual(mockDimensions.width);
-                expect(y).toBeGreaterThanOrEqual(0);
-                expect(y).toBeLessThanOrEqual(mockDimensions.height);
-            });
-        });
-    });
+//             // Verify stone positions are within canvas bounds
+//             positions.forEach((position: number[]) => {
+//                 const [x, y] = position;
+//                 expect(x).toBeGreaterThanOrEqual(0);
+//                 expect(x).toBeLessThanOrEqual(mockDimensions.width);
+//                 expect(y).toBeGreaterThanOrEqual(0);
+//                 expect(y).toBeLessThanOrEqual(mockDimensions.height);
+//             });
+//         });
+//     });
 
-    describe('Responsive Layout Handling', () => {
-        it('should adjust coordinates for narrow screens', () => {
-            const narrowScreenPositions = createStonePositionsDAO({
-                ...mockDimensions,
-                width: 400,
-                riveCanvasWidth: 400
-            });
+//     describe('Responsive Layout Handling', () => {
+//         it('should adjust coordinates for narrow screens', () => {
+//             const narrowScreenPositions = createStonePositionsDAO({
+//                 ...mockDimensions,
+//                 width: 400,
+//                 riveCanvasWidth: 400
+//             });
 
-            const wideScreenPositions = createStonePositionsDAO(mockDimensions);
+//             const wideScreenPositions = createStonePositionsDAO(mockDimensions);
 
-            // Verify that stones are positioned differently on narrow screens
-            expect(narrowScreenPositions[0][0]).not.toBe(wideScreenPositions[0][0]);
-        });
+//             // Verify that stones are positioned differently on narrow screens
+//             expect(narrowScreenPositions[0][0]).not.toBe(wideScreenPositions[0][0]);
+//         });
 
-        it('should use fallback dimensions when Rive canvas dimensions are not set', () => {
-            const positionsWithoutRive = createStonePositionsDAO({
-                ...mockDimensions,
-                riveCanvasWidth: undefined,
-                riveCanvasHeight: undefined
-            });
+//         it('should use fallback dimensions when Rive canvas dimensions are not set', () => {
+//             const positionsWithoutRive = createStonePositionsDAO({
+//                 ...mockDimensions,
+//                 riveCanvasWidth: undefined,
+//                 riveCanvasHeight: undefined
+//             });
 
-            // Verify we still get valid positions
-            expect(positionsWithoutRive).toHaveLength(8);
-            positionsWithoutRive.forEach((position: number[]) => {
-                expect(position[0]).toBeDefined(); // x coordinate
-                expect(position[1]).toBeDefined(); // y coordinate
-            });
-        });
-    });
+//             // Verify we still get valid positions
+//             expect(positionsWithoutRive).toHaveLength(8);
+//             positionsWithoutRive.forEach((position: number[]) => {
+//                 expect(position[0]).toBeDefined(); // x coordinate
+//                 expect(position[1]).toBeDefined(); // y coordinate
+//             });
+//         });
+//     });
 
-    describe('Stone Positioning', () => {
-        it('should maintain minimum distance between stones', () => {
-            const positions = createStonePositionsDAO(mockDimensions);
+//     describe('Stone Positioning', () => {
+//         it('should maintain minimum distance between stones', () => {
+//             const positions = createStonePositionsDAO(mockDimensions);
             
-            // Check minimum distance between any two stones
-            const minDistance = 50; // Minimum acceptable distance in pixels
+//             // Check minimum distance between any two stones
+//             const minDistance = 50; // Minimum acceptable distance in pixels
             
-            for (let i = 0; i < positions.length; i++) {
-                for (let j = i + 1; j < positions.length; j++) {
-                    const distance = Math.sqrt(
-                        Math.pow(positions[i][0] - positions[j][0], 2) +
-                        Math.pow(positions[i][1] - positions[j][1], 2)
-                    );
-                    expect(distance).toBeGreaterThan(minDistance);
-                }
-            }
-        });
+//             for (let i = 0; i < positions.length; i++) {
+//                 for (let j = i + 1; j < positions.length; j++) {
+//                     const distance = Math.sqrt(
+//                         Math.pow(positions[i][0] - positions[j][0], 2) +
+//                         Math.pow(positions[i][1] - positions[j][1], 2)
+//                     );
+//                     expect(distance).toBeGreaterThan(minDistance);
+//                 }
+//             }
+//         });
 
-        it('should position middle stone correctly', () => {
-            const positions = createStonePositionsDAO(mockDimensions);
-            const middleStone = positions[4]; // Middle stone index
+//         it('should position middle stone correctly', () => {
+//             const positions = createStonePositionsDAO(mockDimensions);
+//             const middleStone = positions[4]; // Middle stone index
             
-            // Middle stone should be within reasonable bounds
-            const [x, y] = middleStone;
-            expect(x).toBeGreaterThan(mockDimensions.width * 0.1);  
-            expect(x).toBeLessThan(mockDimensions.width * 0.5);     
-            expect(y).toBeGreaterThan(mockDimensions.height * 0.1);  
-        });
-    });
-});
+//             // Middle stone should be within reasonable bounds
+//             const [x, y] = middleStone;
+//             expect(x).toBeGreaterThan(mockDimensions.width * 0.1);  
+//             expect(x).toBeLessThan(mockDimensions.width * 0.5);     
+//             expect(y).toBeGreaterThan(mockDimensions.height * 0.1);  
+//         });
+//     });
+// });
