+++ src/gameStateService/game-state-service.ts
@@ -2,7 +2,6 @@ import { PubSub } from '../events/pub-sub-events';
 import { DataModal } from "@data";
 import {
     createGameplaySceneDAO,
-    createStonePositionsDAO,
     createLevelEndDataDAO
 } from './data-access-objects';
 
@@ -23,15 +22,6 @@ export class GameStateService extends PubSub {
         LEVEL_END_DATA_EVENT: string;
     }
     public data: null | DataModal;
-    public canvas: null | HTMLCanvasElement;
-    public width: number;
-    public height: number;
-    public canavsElement: null | HTMLCanvasElement;
-    public context: null | CanvasRenderingContext2D;
-    public gameCanvasContext: null | CanvasRenderingContext2D;
-    public riveCanvas: null | CanvasRenderingContext2D;
-    public riveCanvasWidth: number;
-    public riveCanvasHeight: number;
     public isGamePaused: boolean;
     public gamePlayData: null | {
         currentLevelData: {
@@ -70,7 +60,7 @@ export class GameStateService extends PubSub {
         great: string
     };
     //public clickTrailToggle: boolean;
-    public offsetCoordinateValue: number;
+    //public offsetCoordinateValue: number;
     public levelEndData: null | {
         starCount: number,
         currentLevel: number,
@@ -86,23 +76,6 @@ export class GameStateService extends PubSub {
             LEVEL_END_DATA_EVENT: 'LEVEL_END_DATA_EVENT' // To move this event on DOM Event once created.
         };
         this.data = null;
-        /*
-            These need to be moved from the game state to the game settings, some to DOM events for better organization.
-            To Do: Transfer Canvas States variables to game settings
-            To Do: Transfer some Gameplay States - feedbackAudios, feedbackTexts, majVersion, minVersion, and offsetCoordinateValue to game settings.
-            To Do: Transfer clickTrailToggle and updateGameTrailToggle to DOM Events class.
-            To DO: Transfer createLoadingSceneDAO to game settings.
-        */
-        /* Start of Canvas States */
-        this.canvas = null;
-        this.width = 0;
-        this.height = 0;
-        this.canavsElement = null;
-        this.context = null;
-        this.gameCanvasContext = null;
-        this.riveCanvas = null;
-        this.riveCanvasWidth = 0;
-        this.riveCanvasHeight = 0;
         /* Gameplay States */
         this.isGamePaused = false;
         this.gamePlayData = null;
@@ -111,7 +84,6 @@ export class GameStateService extends PubSub {
         this.rightToLeft = false;
         this.majVersion = 0;
         this.minVersion = 0;
-        this.offsetCoordinateValue = 32; //Default value used to offset stone coordinates.
         this.initListeners();
 
         this.levelEndData = null
@@ -135,21 +107,9 @@ export class GameStateService extends PubSub {
         this.isGamePaused = isPaused;
     }
 
-    setDefaultGameStateValues(
-        data: DataModal,
-        canvas: HTMLCanvasElement,
-        canavsElement: HTMLCanvasElement
-    ) {
+    setDefaultGameStateValues(data: DataModal) {
         /*Original game data from FeedTheMonster.ts.*/
         this.data = data;
-        /*HTML and Canvas state values.*/
-        this.canvas = canvas; 
-        this.width = canvas.width;
-        this.height = canvas.height;
-        this.canavsElement = canavsElement;
-        this.context = canavsElement.getContext("2d");
-        this.gameCanvasContext = canvas.getContext("2d", { willReadFrequently: true });
-        /*Gameplay Scene Data */
         this.feedbackTexts = data.FeedbackTexts;
         this.feedbackAudios = data.FeedbackAudios;
         this.rightToLeft = data.rightToLeft;
@@ -161,9 +121,6 @@ export class GameStateService extends PubSub {
         return createGameplaySceneDAO(this);
     }
 
-    getStonePositions() {
-        return createStonePositionsDAO(this);
-    }
     // TODO: move this back to level end scene since 
     levelEndSceneData({levelEndData, data}) {
         this.levelEndData = {...levelEndData};
